requirements.txt
asn1crypto==0.24.0
backcall==0.1.0
cffi==1.12.3
cryptography==2.7
decorator==4.4.0
ipdb==0.12
ipython==7.6.0
ipython-genutils==0.2.0
jedi==0.14.0
jwt==0.6.1
paho-mqtt==1.4.0
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
PyJWT==1.4.0
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
